Move remote bluetooth remote scanner implementation into a base class
- The same implementation can be reused for other remotes
  such as shelly in #82007
